requirements.txt
google-generativeai>=0.3.0
flask>=2.0.0
python-dotenv>=0.19.0
matplotlib>=3.4.0
pandas>=1.3.0
numpy>=1.20.0
plotly>=5.10.0
